feat(account): move the env-wide share list to the account page (AC-11)
Owner feedback: the deployments list belongs with the account surface,
not Preferences. The per-env `fused share list` view (with per-mount
Revoke) moves to a Deployments section on the Fused account page beside
the environments table; Preferences keeps only the Deploy-button toggle
plus a link. The list is the extracted DeploymentsList component,
behavior unchanged.

While moving it, the account page's Environments and Deployments
sections now render in BOTH auth states: the env store and an AWS env's
share list need the CLI, not a managed-Fused sign-in — an AWS-only user
must not pass through an irrelevant sign-in to revoke a link. Only the
account summary and the setup panel stay gated on logged_in.

SPEC: new AC-11; DP-2/DP-13 consumers updated; PF-6 marked moved.
README and code comments updated to match. Verified in the browser in
both auth states (share row + Revoke render signed out and signed in).
